Remove cache
There is no real load on my server, so why not to create some problems for
myself? Actually no, I just want to add a feature for the project, as it
lacks those.

I also didn't fully remove cache, it is reduced to 3 seconds, as I
remember, that some code do something with recursivity or some other
anti-pattern in its case, counting on caching, so it would not do
CPU brrrrr.